TorExitNodes/src/requirements.txt
certifi==2021.10.8
pycti==5.1.3
python_dateutil==2.8.2
PyYAML==6.0
stix2==3.0.1
